tasks: Full migration scripts now deploys independent markets. Added new task to attach a new market to the LendingPoolAddressesProviderRegistry.
