+++ package.json
@@ -40,16 +40,23 @@
     "test:main:check-list": "MAINNET_FORK=true TS_NODE_TRANSPILE_ONLY=1 hardhat test  test-suites/test-aave/__setup.spec.ts test-suites/test-aave/mainnet/check-list.spec.ts",
     "dev:coverage": "buidler compile --force && buidler coverage --network coverage",
     "aave:evm:dev:migration": "npm run compile && hardhat aave:dev",
-    "aave:docker:full:migration": "npm run compile && npm run hardhat:docker -- aave:mainnet --deploy-registry",
-    "aave:kovan:full:migration": "npm run compile && npm run hardhat:kovan -- aave:mainnet --verify --deploy-registry",
-    "matic:mumbai:full:migration": "npm run compile && npm run hardhat:mumbai matic:mainnet --deploy-registry",
-    "matic:matic:full:migration": "npm run compile && npm run hardhat:matic matic:mainnet --deploy-registry",
-    "amm:kovan:full:migration": "npm run compile && npm run hardhat:kovan -- amm:mainnet --verify --deploy-registry",
-    "aave:docker:full:migration:skip-registry": "npm run compile && npm run hardhat:docker -- aave:mainnet",
-    "aave:kovan:full:migration:skip-registry": "npm run compile && npm run hardhat:kovan -- aave:mainnet --verify",
-    "matic:mumbai:full:migration:skip-registry": "npm run compile && npm run hardhat:mumbai matic:mainnet",
-    "matic:matic:full:migration:skip-registry": "npm run compile && npm run hardhat:matic matic:mainnet",
-    "amm:kovan:full:migration:skip-registry": "npm run compile && npm run hardhat:kovan -- amm:mainnet --verify",
+    "aave:docker:full:migration": "npm run compile && npm run hardhat:docker -- aave:mainnet",
+    "aave:kovan:full:migration": "npm run compile && npm run hardhat:kovan -- aave:mainnet",
+    "matic:mumbai:full:migration": "npm run compile && npm run hardhat:mumbai matic:mainnet",
+    "matic:matic:full:migration": "npm run compile && npm run hardhat:matic matic:mainnet",
+    "amm:kovan:full:migration": "npm run compile && npm run hardhat:kovan -- amm:mainnet",
+    "aave:docker:add-market-to-registry-from-config": "npm run compile && npm run hardhat:docker -- add-market-to-registry --pool Aave",
+    "aave:kovan:add-market-to-registry-from-config": "npm run compile && npm run hardhat:kovan -- add-market-to-registry --pool Aave",
+    "matic:mumbai:add-market-to-registry-from-config": "npm run compile && npm run hardhat:mumbai add-market-to-registry --pool Matic",
+    "amm:kovan:add-market-to-registry-from-config": "npm run compile && npm run hardhat:kovan -- add-market-to-registry --pool Amm",
+    "matic:matic:add-market-to-registry-from-config": "npm run compile && npm run hardhat:matic add-market-to-registry --pool Matic",
+    "aave:main:add-market-to-registry-from-config": "npm run compile && npm run hardhat:main -- add-market-to-registry --pool Aave",
+    "aave:docker:add-market-to-new-registry": "npm run compile && npm run hardhat:docker -- add-market-to-registry --pool Aave --deploy-registry",
+    "aave:kovan:add-market-to-new-registry": "npm run compile && npm run hardhat:kovan -- add-market-to-registry --pool Aave --verify --deploy-registry",
+    "matic:mumbai:add-market-to-new-registry": "npm run compile && npm run hardhat:mumbai add-market-to-registry --pool Matic --verify --deploy-registry",
+    "amm:kovan:add-market-to-new-registry": "npm run compile && npm run hardhat:kovan -- add-market-to-registry --pool Amm --verify --deploy-registry",
+    "matic:matic:add-market-to-new-registry": "npm run compile && npm run hardhat:matic -- add-market-to-registry --pool Matic --verify --deploy-registry",
+    "aave:main:add-market-to-new-registry": "npm run compile && npm run hardhat:matic -- add-market-to-registry --pool Matic --verify --deploy-registry",
     "aave:kovan:full:initialize": "npm run hardhat:kovan -- full:initialize-lending-pool --verify --pool Aave",
     "aave:ropsten:full:migration": "npm run compile && npm run hardhat:ropsten -- aave:mainnet --verify",
     "aave:fork:main:tenderly": "npm run compile && npm run hardhat:tenderly-main -- aave:mainnet",
